Remove regression tests and update performance workflow
