--- .github/workflows/Regression.yml
@@ -1,43 +1,0 @@
-name: Regression
-description: "last 3 versions of the library"
-
-on:
-  #pull_request:
-  schedule:
-    - cron: "25 */2 * * *" # Runs every 2 hours at 25 minutes past
-  workflow_dispatch:
-
-jobs:
-  test:
-    timeout-minutes: 10
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        test: [functional]
-        env: [dev, production]
-    env:
-      DATADOG_API_KEY: ${{ secrets.DATADOG_API_KEY }}
-      REGION: ${{ vars.REGION }}
-      SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
-    steps:
-      - uses: actions/checkout@v4
-      - name: Setup test env
-        uses: ./.github/actions/xmtp-test-setup
-
-      - name: Run tests
-        run: yarn test ${{ matrix.test }} --sync all --versions 3 --no-fail --log warn --file
-
-      - name: Cleanup and upload artifacts
-        if: always()
-        uses: ./.github/actions/xmtp-test-cleanup
-        with:
-          test-name: ${{ matrix.test }}
-          env: ${{ matrix.env }}
-
-      - name: Send Slack notification on failure
-        if: failure() || cancelled()
-        uses: ./.github/actions/slack-notification
-        with:
-          workflow-name: "Regression (${{ matrix.test }}, ${{ matrix.env }})"
-          slack-bot-token: ${{ secrets.SLACK_BOT_TOKEN }}
