+++ package.json
@@ -23,7 +23,7 @@
     "monitor:yarn": "httpstat https://grpc.dev.xmtp.network:443",
     "permissions": "tsx cli/permissions.ts",
     "record": "npx playwright codegen 'https://xmtp.chat/'",
-    "regression": "yarn test functional --env local --sync all --versions 3 --size 10-50-100",
+    "regression": "yarn test performance --env local --sync all --versions 3 --size 10-50-100",
     "revoke": "tsx cli/revoke.ts",
     "send": "tsx cli/send.ts",
     "test": "tsx cli/test.ts",
